Cleanup in api.go
