+++ pkg/keboola/storage_common.go
@@ -1,0 +1,41 @@
+package keboola
+
+type Object interface {
+	ObjectID() any
+}
+
+// Metadata - object metadata.
+type Metadata map[string]string
+
+// MetadataDetails - metadata with details (id, timestamp).
+type MetadataDetails []MetadataDetail
+
+// MetadataDetail - metadata with details (id, timestamp).
+type MetadataDetail struct {
+	ID        string `json:"id"`
+	Key       string `json:"key"`
+	Value     string `json:"value"`
+	Timestamp string `json:"timestamp"`
+}
+
+// ToMap converts slice to map.
+func (v MetadataDetails) ToMap() Metadata {
+	out := make(Metadata)
+	for _, item := range v {
+		out[item.Key] = item.Value
+	}
+	return out
+}
+
+// DeleteOption for requests to delete bucket or table.
+type DeleteOption func(c *deleteConfig)
+
+type deleteConfig struct {
+	force bool
+}
+
+func WithForce() DeleteOption {
+	return func(c *deleteConfig) {
+		c.force = true
+	}
+}
